requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.62.0

# Transcript processing
youtube-transcript-api>=0.4.4
nltk>=3.6.0
spacy>=3.2.0

# Text-to-Speech
soundfile>=0.10.3
kokoro>=0.1.1

# AI processing
openai>=0.27.0
langchain>=0.0.200
litellm>=1.0.0
google-generativeai>=0.3.0

# Web and API
Flask>=2.0.1
flask-cors>=3.0.10

# Utility libraries
requests>=2.26.0
python-dotenv>=0.19.0

# Development dependencies
pytest>=6.2.5
black>=21.9b0
isort>=5.9.3
